Creation of Details Service

+++ API_Shopping/API_Shopping/Models/Detail.cs
@@ -4,8 +4,8 @@
     {
         public long ?Id { get; set; }
         public int Quantity{ get; set; }
-        public int Price{ get; set; }
-        public int Total { get; set;  }
+        public decimal Price{ get; set; }
+        public decimal Total { get; set;  }
         public DateTime? UpdateAt { get; set; }
 
         public long OrderId { get; set; }
